add new response class

+++ app/src/main/java/by/isb/an07/hw8/data/entities/crypto/Crypto.kt
@@ -1,19 +1,19 @@
 package by.isb.an07.hw8.data.entities.crypto
 
 data class Crypto(
-    val circulatingSupply: Int,
+    val circulatingSupply: Double,
     val cmcRank: Int,
     val dateAdded: String,
     val id: Int,
     val lastUpdated: String,
-    val maxSupply: Int,
+    val maxSupply: Double,
     val name: String,
     val numMarketPairs: Int,
 //    val platform: Any,
     val slug: String,
     val symbol: String,
 //    val tags: List<String>,
-    val totalSupply: Int,
+    val totalSupply: Double,
 //    val marketCap: Double,
     val percentChange1h: Double,
     val percentChange24h: Double,
